demo-schedule rename to demo-reports

+++ demo-admin/src/main/java/com/demo/admin/entity/MenuStructure.java
@@ -12,5 +12,5 @@ public class MenuStructure extends BaseEntity {
     private Long parent;
     private String icon;
     private String name;
-    private Long roleId;
+    private Integer roleId;
 }
